msm: camera: Fix early camera initialization to disable rotary switch
Fix initialization of early camera to take into account
state of all switches to determine if it should be on or
off. Rotary switch is disabled for rearview camera as
some platforms may not have correct firmware to have gpio
appropriately setup for boot. Delay adv start to ensure
coherent frame.

CRs-fixed: 827172
Change-Id: I4ef4fde7b4ab65b0f04f3306553fa93ab69fb4bb
Signed-off-by: Shiju Mathew <shijum@codeaurora.org>