msm: mdss: prevent deadlock while enabling/disabling hist irq
While enabling or disabling histogram irq, first the spinlock
hist_lock is acquired and then ultimately a spinlock mdp_lock
is acquired before programming registers. Whereas when irq
handlers are dispatched the spinlock mdp_lock is first acquired
and in hist done interrupt handler the spinlock hist_lock is
attempted again for acquiring. This causes a deadlock scenario
which can cause watchdog timeout.

Histogram interrupt registers need not be protected by mdp_lock
as the address for the registers is different and they are
protected with spinlock of hist_intr. Hence removing the
mdp_lock protection to the code for writing to histogram irq
registers.

Change-Id: I860639c81789a4080344610071834afa42c8eb4b
Signed-off-by: Krishna Chaitanya Parimi <cparimi@codeaurora.org>