msm: mdss: update the pp status flags for layermixer
Post Processing(PP) driver caches the status of the features enabled.
Information is used to turn on the PP features when device is resumed
from suspend state. Currently PP status information is not updated for
layer mixer path. Add support to track the layer mixer status
and prevent programming the DSPP blocks if dirty flag is not set.

Change-Id: I5fa83e418f648a74ee5bf17285d3a9f5615fbd69
Signed-off-by: Yexuan Zhang <lyzhang@codeaurora.org>