diag: Set the diag smd buffer to in busy on receiving the SMD interrupt
Currently the in_busy state of SMD buffer is set only after reading
the data present in the command response channel or if the SMD buffer
is full. This can cause buffer corruption if we get another SMD
interrupt in between. Make a change to set the in_busy state before
reading the data into the SMD buffer.

Change-Id: Ic4b1ff1bc5134d26e89867e0ee63829094fd3b90
Signed-off-by: Sreelakshmi Gownipalli <sgownipa@codeaurora.org>