usb: dwc3_otg: Add support for driving CPE USB SW pin
CPE USB SW pin on MDM9x40 MTP debug board is required to be
driven using GPIO control. This is required to interface
with a second debug board that routes USB connection to PC
(when CPE USB SW is LOW) or to USB-to-Ethernet
(when CPE USB SW HIGH). When operating in host mode drive
the CPE USB SW PIN HIGH using GPIO control so that the path
to USB-to-ethernet is active.

Change-Id: I93049715f2e3ef9a7ab4af34a1c437edab4edfa8
Signed-off-by: Devdutt Patnaik <dpatnaik@codeaurora.org>