ARM: dts: msm: Set minimum CPUBW based on CPU freq for MSM8952
This allows setting a DDR frequency floor based on CPU frequency is
addition to using the bw_hwmon governor for decide the final CPU BW
vote. Because of how the BW votes are aggregated, if bw_hwmon pick a lower
vote than the floor, the final vote will be the floor value.

This patch also fixes rounding of IB values. IB values should always
round down.

Change-Id: I710e3f02b1c06ad0defa120e98d8099334eb2aaa
Signed-off-by: Arun KS <arunks@codeaurora.org>
Signed-off-by: Hanumath Prasad <hpprasad@codeaurora.org>