msm: mdss: tie histogram collection to commit with irq mask/unmask
Histogram collection needs to be tied to commit. Clients utilizing
histogram collection would program the PA LUT after software
processing of the histogram. The PA LUT would only take effect
over a CTL flush, which needs to be done over a commit call. So
histogram collection is now also tied to commit.

Histogram irq is enabled and disabled as per following:
1. Histogram irq enablement and disablement to tie it to commit
   is only done if that particular histogram block was first
   enabled using call to start histogram.
2. Histogram irq is set during commit to get the interrupt.
3. Histogram irq is disabled during the second consecutive
   interrupt handler call for a particular block without any
   commit call in between the two interrupt handler calls. This
   is due to the following.

Fig: Timeline of vsyncs, commits and histogram collections
               h0           h1           h2           h3
|------:-----|-^----:-----|-^----------|-^----:-----|-x----
v0     c1    v1     c2    v2           v3     c4    v4
 <----f0----> <----f1----> <----f2----> <----f3----> <--f4-

v0 - v4 are vsyncs,
c1 - c4 are commits,
h0 - h3 are histogram collections
f0 - f4 are frames being fetched in pipes  over the timeline.

The frame f0 is being fetched through the pipe after v0 till v1,
because of which histogram for it is collected at h0 after v1.
Similarly for frame f1 the histogram h1 is collected after v2.

If we were to disable histogram irq right after histogram interrupt
say h0, then we will not be able to re-enable it till v2 (next
vsync in timeline). So we would not be able to collect the h1
histogram for frame f1 being fetched during v1 to v2. So we will
be collecting alternative frames at best.

If instead we disable histogram irq after two consecutive
interrupts without a commit in between then we would not miss h1,
and would disable after h2. After h2, c4 would arrive but that
re-enables irqs at v4. However the frame being fetched during
this time (from v3 to v4) is f3, which is same as f2 as no commit
came in between. So we would miss out on h3 for f3 which is
correctly omitted.

Change-Id: I75be1651a3842f004c07f2ce8a5ea2bed476b64e
Signed-off-by: Krishna Chaitanya Parimi <cparimi@codeaurora.org>