mm: process_reclaim: search for mm in the thread group
The reclaim function picks the mm of the task_struct passed
to it, and skips the reclaim if mm is NULL. This does not take
into consideration the case of, tasks which have detached the mm
but still other tasks in the thread group using the same.

Find the mm by iterating over the thread group. This requires the
the thread group to be intact during the process, in turn fixing an
exisitng problem where test_task_flag called from swap_fn can result
in a crash because of the thread_group elements being removed in
parallel.

Change-Id: I36269b1b6619315f33f6f3b49ec73571a66796f2
Signed-off-by: Vinayak Menon <vinmenon@codeaurora.org>