power: qpnp-fg: fix error in delayed_init_work
Commit c4301419358485fa265df9c1adac9ec1a600418c ("power:
qpnp-fg: reduce the probe time of driver) moved the hardware
initialization to delayed_init_work to improve the driver probe
time. However, with the recent change made in update_sram_data
to wait for SRAM access to be revoked before updating SRAM
parameters times out because the access is held by the callee
itself.

This eventually causes error in fg_batt_profile_init as below.

[    9.547366] FG: fg_hw_init: set default value to esr filter
[   12.542896] FG: update_sram_data_work: transaction timed out ret=0
[   17.539999] FG: fg_batt_profile_init: profile loading timed out rc=-110
[   17.545583] FG: batt_profile_init: failed to initialize profile

Fix this.

CRs-Fixed: 844197
Change-Id: I16f52f991ccccc275a945874bfe86259a5ec97f1
Signed-off-by: Subbaraman Narayanamurthy <subbaram@codeaurora.org>
Signed-off-by: Xiaozhe Shi <xiaozhes@codeaurora.org>