ASoC: msm8952: parse wsa codec name in 8952 machine driver
WSA codec names are specified in device tree. Parse them in machine
driver so that the sound card can be registered successfully via
mapping the aux device name with the wsa codec.
Also uses wsa counters to match the device tree configuration.
If the matching fails, we assume that there is no wsa881x on board.
It will double ensure the wsa881x existence

Change-Id: I6f6c5909ba82421d02ca6bb1c2b5524f87ee0e95
CRs-Fixed: 841458
Signed-off-by: Walter Yang <yandongy@codeaurora.org>