msm: kgsl: Don't WARN in _kgsl_context_get
There is a sizable window between the reference count on a context
going to zero and it actually getting removed from the idr. In the
meantime somebody else could  try to get the context and fail
kref_get_unless zero(). Thats okay - it just means that you can't
trust the pointer. This mostly comes into play with kgsl_release()
which races with all the other things that hold a context reference.

Remove the WARN and trust that all callers of _kgl_context_get()
do the right thing (as of today, they do).

Change-Id: Ic0dedbadbd9f257042da9ea7e18e7ba4f27c1a0a
Signed-off-by: Jordan Crouse <jcrouse@codeaurora.org>
Signed-off-by: Sunil Khatri <sunilkh@codeaurora.org>