mtd: ubi: fix sync erase fallback in ubi_wl_scrub_all
Before completing the 'scrub all' few PEBs are erased synchronously to
prevent stressing due to not enough free PEBs.
When sync erase fails in ubi_wl_scrub_all a delayed erase is scheduled.
After the delayed erase is scheduled, the wl entry is added to the free
PEB tree. This is wrong, as the erase worker will do that after the PEB
is erased. In addition, if both the sync erase and delayed erase fail,
move fs to read only mode to keep consistency in error handling path
with general ubi code.
This also fixes the implicit dependency of ubi_wl_scrub_all with
fastmap in compilation.

Change-Id: Icd7c79775f40085026cc785234d78db67d705fae
Signed-off-by: Dolev Raviv <draviv@codeaurora.org>
Signed-off-by: Tanya Brokhman <tlinder@codeaurora.org>