USB: gadget: u_bam_data: Use gadget's global function workqueue
In very fast back to back suspend resume scenario, there is a race
happening when first suspend work is still running we get resume
followed by suspend again.
As second suspend has happened we see that A_BUS_REQ is clear, when
trying to enter LPM and USB is put in LPM. Afterwards when resume
work is queued and due to unclocked access causes target to crash.

Avoid the above mentioned case by using gadget's function workqueue, which
is flushed as part of msm_otg suspend before going to LPM.

Change-Id: I0be72e30e69f1eea9b9fa6b5117bd77c00c2ab0b
Signed-off-by: Manu Gautam <mgautam@codeaurora.org>