msm: ipa: invalidate uC requested command after completion
This change invalidates the command that is being sent to
IPA uC once a completion is received (whether it's in polling or
interrupt mode). This is done in order to prevent unecessary calls
for the completion object of the IPA uC after a response has
already been received by polling (As the interrupt is fired in
any case).

Change-Id: Ifa7531324e9c9a6b2e0714f6e043d56eb534c4b8
Acked-by: David Arinzon <darinzon@qti.qualcomm.com>
Signed-off-by: Sivan Reinstein <sivanr@codeaurora.org>