clk: qcom: cpu-clock-8939: Add support for pcnoc vote
Whenever APSS is active, the cpu clocks would need to put an active vote
of 50MHz on peripheral NOC. This would in turn put a vote on GPLL0, which
would help in cases where APPS needs to be out of suspend and gpll0 is off
due to no other subsystem vote.

Also add a device tree property to indicate if this vote is required or
not.

Change-Id: I531480f13013cd89d3866851f8f3bf591f8dc41b
Signed-off-by: Taniya Das <tdas@codeaurora.org>