qcom: cpu_ops: dump pending interrupts upon hotplug crash
In rare cases, the hotplug out operation seems to fail and
hits the BUG() in msm_wfi_cpu_die. One possible reason
could be a pending interrupt.Dump all the pending interrupt
data just before BUG() to find out the cause.
This patch is experimental to debug a corner case that is
very difficult to reproduce and it can be reverted back later.
Further, since the dump is done just before BUG(),
there are absolutely no side effects of this patch.

Change-Id: I271a72547b6357a347f31be3be85d2af369273ac
Signed-off-by: Lingutla Chandrasekhar <clingutla@codeaurora.org>