regulator: qpnp-labibb-regulator: fix configuring MODULE_RDY of LAB and IBB
Currently, MODULE_RDY of both LAB and IBB are getting configured
in HLOS only if the LAB/IBB modules are configured already by the
bootloader. However, if the LAB/IBB regulators needs to be
configured and used from the HLOS, MODULE_RDY should get
configured in HLOS if not already. Modify the configuration of
those registers accordingly.

CRs-Fixed: 877711
Change-Id: I17b6eea1b6ade2b76870af1833270af22af842f5
Signed-off-by: Subbaraman Narayanamurthy <subbaram@codeaurora.org>