msm: mdss: reset backlight scale variables on framework reboot
On framework reboot the backlight scale variables are not reset.
Hence after the framework reboot when backlight altering modules
restart there can be a very likely scenario for a flicker. Hence
reset the scale variables prior to shutdown. This would allow
the variables to be in reset state if case is a framework reboot
and not shutdown.

Change-Id: I9796c8b7abc34daca4ce0ee0ff1641517fb3d06f
Signed-off-by: Krishna Chaitanya Parimi <cparimi@codeaurora.org>