msm: mdss: always assign virtual address to cursor buffer
Cursor IOCTL is used to program the cursor pipes and the updates
would take effect on display_commit along with the other layers.
In cmd mode panels, when the there are no updates there is idle
power collapse and iommu is detached. If cursor IOCTL is called
during this time, it would assign the physical address instead of
virtual address for cursor buffer. This causes dirty region around
the cursor in the next display_commit. Fix the cursor buffer to have
the virtual address all the time.

Change-Id: I963794d0de5387b10e06e7c5a61300cd7b4aeffa
Signed-off-by: Veera Sundaram Sankaran <veeras@codeaurora.org>