msm: mdss: use touch event trigger to update fps
During idle fallback, mdss driver will configure min fps
to the panel and in the next composition cycle it will
reset it to the requested fps, this will result in latency
for first frame update.

For the usecase, where touch triggers the update, this
latency can be alleviated by updating the fps during touch
so that there is no extra wait during commit.

Change-Id: I625785d27da0b485877488e4586f6a9336081a47
Signed-off-by: Kalyan Thota <kalyant@codeaurora.org>