msm: camera: add support for supported hw streams
We need to have more than one hw stream for one user stream,
but only one of them will output buffer for user stream.
The other hw stream will be used for internal use and using
of native buffers for those streams is mandatory. Because of
that we are going to use only 12 less significant bits to
store user stream id. High significan bits will identify the
hw streams in case of we have more than one hw stream for
one user stream.

Change-Id: I5bf0e033b8ab2b0e229950ac8e4d3cd7471bce3e
Signed-off-by: Vasko Kalanoski <vaskok@codeaurora.org>
Signed-off-by: Konstantin Motov <kmotov@codeaurora.org>