ARM: dts: msm: increase the disconnect threshold for MSM8952
In some board designs, the transmit amplitude of the PHY is increased
to counter the attenuation resulted due to the additional components
between PHY and the connector. If the default disconnect threshold
is used, PHY may report false disconnect event. Hence increase the
disconnect threshold by 6% of the default value.

Change-Id: Ibb0117a32f1200ea3f86f6fff3688cb4aff7f61c
Signed-off-by: Vijayavardhan Vennapusa <vvreddy@codeaurora.org>