Revert "USB: hid: hid-core: Enable autosuspend feature for HID devices"
Commit 3af0f9d718b63d4ec1ca93754ef326929002f5e7 added to allow autosuspend
for HID devices. with this change we have observed suspend/resume issues
with some of the HID devices available in the market. In current design
we removed wakelocks to allow system suspend on ADP platform and no need
to support autosuspend for usb devices.

This change is used to disable runtime suspend for HID devices.

Change-Id: I6a9fd2eee2ce25120569005131ccc763f08884e2
Signed-off-by: ChandanaKishori Chiluveru <cchilu@codeaurora.org>