msm_serial_hs: Remove force stale operation during baud rate configuration
During baud rate configuration there may or may not
be an active Rx tranfer.
If there is no active transfer stale interrupt won't
be triggered for force stale command. So logic ends up
with timeout and results in error.

Instead of force stale command clear the FIFO and
discard the queued Rx commands to ADM.

Change-Id: I58dfbc5bc32eb51928b4c21a50601a26c0db6fcd
Signed-off-by: Dilip Kota <dkota@codeaurora.org>