power: qpnp-smbcharger: avoid suspending charging after max AICL irqs
Currently after we receive 25 AICL interrupts within 60 seconds
we declare the charger to be very weak and disable AICL rerun.
If an AICL event happens again, it will suspend charging until
the charger is removed. AICL lowers the input current to a
minimum of 300mA and it suspends charging if the voltage doesn't
recover.

However, there could be chargers that work fine at 100mA, but we
end up suspending such chargers too.

Change the code to vote for 100mA current limit after declaring a
very weak charger. AICL gets disabled for any input settings
below 300mA. If the charger still collapses, we will get a UV
interrupt and there we will declare the charger as weak and
suspend charging.

CRs-Fixed: 869191
Change-Id: Ic3809fdc11d88538c0dcca43a20a23632c1a2d78
Signed-off-by: Nicholas Troast <ntroast@codeaurora.org>