usb: msm_otg: Disable ID_GND interrupt for OTG user control
In micro-A cable connect/disconnect case, As USB ID line is routed
to PHY on this hardware. An asynchronous USB IRQ generated by the Link
when an ID high voltage interrupt is indicated by the PHY. Which is
causing usb to exit from low power mode on micro-A cable connect/
disconnect on USB OTG port in user control mode.

To fix this disable ID_GND fall and raise interrupt notification
in USB PHY.

Change-Id: Ie6f377d52297c4a84d0bc121d97e7ef5b3c0b867
Signed-off-by: ChandanaKishori Chiluveru <cchilu@codeaurora.org>