power: smb135x-charger: detect HVDCP
The HVDCP algorithm runs after a DCP has been detected by APSD.
Currently we ignore the results of the HVDCP algorithm and
report HVDCP chargers as DCP to the USB power supply.

Read the results of the HVDCP algorithm and notify the USB power
supply if an HVDCP charger has been detected.

CRs-Fixed: 911532
Change-Id: I0d87bc29cd1e8fab9e711f3ea6b72d7d9b66ec67
Signed-off-by: Nicholas Troast <ntroast@codeaurora.org>