msm: sps: avoid register read for ignored debugging messages
When the detailed logging is not enabled, pr_debug() will not
output the logging message. However, the parameters of pr_debug()
are still calculated. Thus, the parameters involving register read
could consume significant CPU time when the function with this kind
of logging message is invoked frequently. The fix here is to ignore
the operation of logging involving register access when the
detailed logging is disabled.

Change-Id: I1aff13c6f054c319034d7a47356bdd345741f159
Signed-off-by: Yan He <yanhe@codeaurora.org>