msm: mdss: Add source crop logic for layer mixer HW cursor support
Source crop for cursor is now sent along with the image width/height.
Previously, cursor was always set with the image with/height as there
was no crop information. Allocate buffer with size of actual incoming
image width and height. This would address the issue of cursor
image being more than the supported cursor size, as previously cursor
buffer was alloted only to a fixed cursor size.

Also, In cmd mode panels, when the there are no updates there is idle
power collapse and iommu is detached. If cursor IOCTL is called
during this time, it would assign the physical address instead of
virtual address for cursor buffer. This causes dirty region around
the cursor in the next display_commit. Fix the cursor buffer to have
the virtual address all the time.

Change-Id: I66e22719a8832f38fe237038925091fbcb02448c
Signed-off-by: Jayant Shekhar <jshekhar@codeaurora.org>