power: qpnp-smbcharger: set 'ALLOW_DETECTION' property at boot-up
It is expected that at charger insertion, ALLOW_DETECTION property be set
prior to setting the usb power supply's PRESENT and TYPE property.

At the time of boot up, if a charger is found present the driver only sets
the PRESENT and TYPE property without setting 'ALLOW_DETECTION' property.
Fix this.

CRs-Fixed: 933626
Change-Id: I09a1939fc7528afe38965c5f561b6ca44bd44b9e
Signed-off-by: Ashay Jaiswal <ashayj@codeaurora.org>