msm: vidc: Check for active buffers only in registered list
In map_and_register_buf(), ensure that the queued buffer is checked
against existing buffers in the registered buffer list which are
marked active. An active buffer is one which is currently in use by
firmware. If the firmware has released all references, then driver
deletes the node from the registered list. If however, the address
range mapped by this buffer node overlaps other buffers in the list
which share the same fd, then the buffer node is not deleted from
the registered list, but marked as inactive.

Change-Id: Ifadfbd8718fa98ed45911a50fa061623bcb16685
Signed-off-by: Arun Menon <avmenon@codeaurora.org>