usb: dwc3-msm: Remove usb_ext notification
The usb_ext notification was intended to allow clients to
receive notification of ID pin state changes. However there
are no current clients, and if and when any are introduced
they should be encouraged to use the extcon framework to be
notified directly of the cable state. Remove the dead code.

Change-Id: Ic7ca8398c6dd4ea1c46a3208791330c17f90e550
Signed-off-by: Jack Pham <jackp@codeaurora.org>