usb: dwc3-msm: Use same workqueue for all work functions
commit 45b16e7bcbee ("usb: dwc3-msm: Use dedicated workqueue for
resume_work") introduced dwc3_wq but is only used for the
resume_work function. Since the idea is to try to synchronize
that function against state machine work function, then sm_work
should also be scheduled on dwc3_wq, as it is created to be an
ordered and single-issued workqueue. Move restart_work there as
well for the same reason. Leave bus_vote_w alone for now since
it isn't critical to be serialized with the rest.

Change-Id: I587dfc88b749f1315f730f084a6f953071d17030
Signed-off-by: Jack Pham <jackp@codeaurora.org>