usb: dwc3-msm: Simplify OTG_STATE_UNDEFINED handling
In commit 48924e2fe107 ("dwc3: Reset USB controller/PHY after psy
connect indication at bootup") the controller initialization path
was removed from dwc3_msm_probe() and is deferred to when actually
needed, which is when beginning peripheral or host mode. This was
to optimize for the HVDCP charger already-connected case in which
case we want to avoid performing controller initialization which
can disrupt D+/D- linestate.

As part of that change, the OTG_STATE_UNDEFINED state was made
to be the entry point for starting controller initialization.
However, apart from that, the handling is identical to B_IDLE_STATE,
and therefore could be consolidated. This patch simplifies the
duplicate code as follows:

- The dwc3_initialize() sequence is nearly identical to that of
  dwc3_msm_resume(), so allowing the state transitions that already
  call pm_runtime_get_sync() achieves the same effect. On the first
  call to resume() perform any necessary one-time initialization
  as needed.
- Setting the DCP current draw can be done immediately in the
  power_supply PROP_PRESENT handling so that scheduling sm_work can
  be avoided entirely.
- This helps to redefine what "B_SESS_VLD" means: cable is connected
  with VBUS and is either an SDP or CDP (but not a DCP charger).
  This is the only time in which start_peripheral() can be called.
- No need to have a completion for initial VBUS notification. The
  state machine will remain in UNDEFINED state until a BSV or ID
  event transitions out of it.
- dwc3_ext_event_notify() can be inlined as part of sm_work() and
  the inputs bitfield is now redundant. Remove them.

Change-Id: I1d53aca38eeb21c3a159b04be792d75875257711
Signed-off-by: Hans Chang <hansc@codeaurora.org>