msm: camera: fix FE violation error
FE finish when ISP processing finish. If we start FE before that we
get violation error. We need to wail until statistics are ready.
It does not guarantee that processing finished, but it is good
enough until ROI of one of statistics cover entire input image.
Statistic frame drop is not allow in case of FE because we rely on
statistics to determinate ISP processing done. AXI done IRQs are not
reliable because we may have FOV crop or frame drop.

Change-Id: I96ac07db4c573ee8b8fed469814401d5a2467be1
Signed-off-by: Konstantin Motov <kmotov@codeaurora.org>