usb: gadget: Clear suspend flag on connect
Driver is setting is_suspended variable to true during suspend and
clearing on resume. When we disconnect the cable and reconnect once
on bootup, there is a suspend interrupt followed by disconnect. In
this case is_suspended variable gets set to TRUE not cleared on cable
connect and causing DUN call/MODEM Query failures.

This change is added to clear the suspended variable during connect.

Change-Id: Id345123a4184e4ffc47f205d5d765f7798db1452
Signed-off-by: ChandanaKishori Chiluveru <cchiluve@codeaurora.org>