ARM: dts: msm: Set WLAN_EN pin to active-low when SDIO is in suspend state
Control the WLAN_EN pin from SDIO driver. Pull it high during SDIO
power-up and pull it down during power-off.
When APPS is in suspend mode and if WiFi is not enabled, WLAN_EN pin
would be set in low power configuration.

Without this change WLAN_EN pin is active all the time and affects the
stand-by current.

Change-Id: Ie7883ea71b76613771caf0a366dad9878a2d6e0c
Signed-off-by: Veerabhadrarao Badiganti <vbadigan@codeaurora.org>