net: Add CNSS2 platform driver
This driver is mainly designed for QCA6290 and afterward PCIe based
Wi-Fi chipset. It is also backward compatible with current QCA6174
PCIe based Wi-Fi chipset. It is leveraged from CNSS platform
driver but has complete code refactoring which is able to provide
more scalability and extensibility for future development. It also
removes duplicated code and unused features from CNSS driver.

This is the initial version of the driver which supports QCA6174 PCIe
based chipset first and adds the hooks for QCA6290 chipset.

Change-Id: I68c6346413dc7bd2a05bef72a6bcf33d76debdc1
Signed-off-by: Yue Ma <yuem@codeaurora.org>
Signed-off-by: Long Cheng <quic_longchen@quicinc.com>