ARM: dts: msm: Support eMMC on sd card slot of MDM9607
The card detect gpio pin of sd card is already used by audio subsystem.
It was earlier decided that the common gpio will be used by audio
subsystem. So disabled the card detection functionality so that eMMC
card and SD card will work in the SD card slot as non removable devices.

When a new eMMC card is inserted for the first time, it needs to be
formatted and partition created so that from next bootup onward, it
will be automatically mounted.

Change-Id: I66b36d04867e53b2df5dcd3507b73869957dbb25
Signed-off-by: Vijay Viswanath <vviswana@codeaurora.org>