msm: Ignore multiple PHY_DL_OVERFLOW csid interrupts
Too many back-to-back PHY_DL_OVERFLOW csid interrupts
on TIF session closure is causing an IRQ storm and a
device reset. This patch disables the corresponding
interrupt mask on the first occurrence of this interrupt.
This is done only in case of TIF use case, where the
first csid device is used.

Change-Id: I7211d8bf03c1eef4337b58c62fb43ed2c10a08c7
Signed-off-by: Rahul Sharma <sharah@codeaurora.org>