usb: f_gps: Fix infinite loop of response available notifications
As part of function suspend we are clearing all the
responses and setting the notify count to 0. If a queued
request on the interrupt endpoint is not
completed before function suspend and is completed
after the next resume, the notify count is decremented
below 0 and causes an infinite loop of response
available notifications. Prevent this by dequeuing any
request on the interrupt endpoint during function suspend.

CRs-Fixed: 2100624
Change-Id: I75425ae84d0901a0ac34c34db3a7bbb527143540
Signed-off-by: Sriharsha Allenki <sallenki@codeaurora.org>