usb: gadget: f_gps: Do not clean up response queue on suspend
The current driver cleans up the response queue on suspend.
This results in data loss. Fix this by not clearing the
responses on suspend and wakeup the host if any response
is available. And the response is also dropped if there
is an enqueuing error on interrupt endpoint which might
lead to data loss. Fix this by not dropping the response
on error.
If the device is suspended and remote wakeup is not
enabled, all the responses will be queued and may lead
to a memory shortage. Fix this by limiting the responses
that can be queued.

Change-Id: I4390602875d3e85f333b107b058e299f973c837c
Signed-off-by: Sriharsha Allenki <sallenki@codeaurora.org>
Signed-off-by: Ajay Agarwal <ajaya@codeaurora.org>