usb: gadget: f_gps: Fix multiple notifications for single response
When there are multiple responses pending with
the driver, the notify count is greater than 1.
In this case host may acknowledge multiple notifications
but is still fetching only one response. This breaks
the sync between the notify count and number of
responses and results in data loss.
Fix this by notifying the host of a available response
only on the completion handler of previous
get response.

Change-Id: Ica48682246ecaacc4cdc1f45edf60ad954636663
Signed-off-by: Sriharsha Allenki <sallenki@codeaurora.org>
Signed-off-by: Ajay Agarwal <ajaya@codeaurora.org>