usb: gadget: f_mbim: Clear notify count on RESET_FUNCTION
Upon reset function command from the host we are clearing
the response queue but the notify count is still non-zero.
This causes to queue a notification to host even if there
are no responses available with this resulting in an
connection failure.
Fix this by clearing the notify count and dequeue any
pending request on the interrupt endpoint.

Signed-off-by: Sriharsha Allenki <sallenki@codeaurora.org>
Change-Id: Ie7e593214b55523a7a25feb08d05665752ebf1f0