thermal: qpnp-adc-tm: Fix aborting suspend
When a low hysteresis for thresholds are set, there is a
possibility that low and high thresholds are crossed at
around the same time before the client is notified. Under
such condition there could be a low and high interrupt
received before processing the request and notifying the
client. Check the return value for queued item for
notifying the client and reduce the count used to prevent
suspend if the item is already available in the global queue.

Change-Id: I2eca945a3dd8d6c5ca07801f8100a5e5c90862b2
Signed-off-by: keqili <keqili@codeaurora.org>