usb: gadget: f_hid: fix: Prevent accessing released memory
commit aa65d11aa008f4de58a9cee7e121666d9d68505e upstream.
When we unlock our spinlock to copy data to user we may get
disabled by USB host and free the whole list of completed out
requests including the one from which we are copying the data
to user memory.
To prevent from this let's remove our working element from
the list and place it back only if there is sth left when we
finish with it.
Fixes: 99c515005857 ("usb: gadget: hidg: register OUT INT endpoint for
SET_REPORT")
Cc: stable@vger.kernel.org
Tested-by: David Lechner <david@lechnology.com>
Bug: 74447444
Signed-off-by: Krzysztof Opasiak <k.opasiak@samsung.com>
Signed-off-by: Felipe Balbi <felipe.balbi@linux.intel.com>
Cc: Jerry Zhang <zhangjerry@google.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
Git-commit: 000bb814a0518c1d3a060ed913b7ecf19316e71a
Git-repo: https://android.googlesource.com/kernel/msm
[srkupp@codeaurora.org: Resolved conflicts]
Signed-off-by: Srinivasa Rao Kuppala <srkupp@codeaurora.org>

Change-Id: Ie5a91a3a2d259a58156de44796815c9ba5cb3eec
Signed-off-by: kaiyiz <kaiyiz@codeaurora.org>