ima: Define IMA appraisal/measure rules based on uuid
Previously appraisal rules were defined in a policy file
in user space. In order to obtain better security, the
rules are now defined in kernel itself to address the
case where policy file itself is compromised due to malware.

When a ubi image is created, uuid is generated which is
a random number every time. This image when flashed in
a partition helps in identifying that partition based
on uuid. Because we need to define a UUID in IMA rule,
we are using pre-defined UUIDs for root and data partitions.

Change-Id: I44b2e3f4d3bc39aa49602fb42dbebaa9a3e1a738
Signed-off-by: Rishi Gupta <rishgupt@codeaurora.org>