ima: disable creating sysfs entries for ima policy
The rules to measure and appraise files are now embedded
in driver file ima-policy.c itself. Therefore there is no
need for sysfs files that are required to send new rules
to kernel. This commit removes call to the logic to create
those files. Additionally, it increases security a bit also.

Change-Id: Ia3bfd0ef600e66807b08209778bbaad0c13b13e1
Signed-off-by: Rishi Gupta <rishgupt@codeaurora.org>