usb: dwc3: Disable device events after maximum error retries
Currently if the core could not recover even after maximum
number of error retries there can storm of error interrupts
generated since the interrupts are not handled because
of dwc->err_evt_seen being set on first error event seen.
Hence after maximum error retries disable the device
events and keep the core in softreset state. Additionally,
reset the err_evt_seen on run/stop clear to ensure
that the interrupts on the next connect/session are
processed properly.

Also, the current check for b_suspend and vbus_active
is incorrectly preventing the suspend as part of restart
which will prevent the core from recovering.
Fix this by adding a in_restart flag to ensure that the
core is recovered and reset as part of restart work.

Change-Id: I760e163733bdcc6230d926fa548aa93a5da918a5
Signed-off-by: Sriharsha Allenki <sallenki@codeaurora.org>