init: define early init functions for rootfs mounting
Here we define init functions for our first subsystem which
initializes flash controller, filesystem, and starts our
early userspace applications.
And ufs scan runs in sync mode if early feature is enabled
so it runs in the thread for rootfs mounting and we do not
need additional complicated code to do the sync job.

Change-Id: I8a81beba52a78b0a7f239e2d0252a69972865b6c
Signed-off-by: Jiacheng Zheng <jiaczhen@codeaurora.org>