quota: avoid time_t in v1_disk_dqblk definition
The time_t type is part of the user interface and not always the
same, with the move to 64-bit timestamps and the difference between
architectures.

Make the quota format definition independent of this type and use
a basic type of the same length. Make it unsigned in the process
to keep the v1 format working until year 2106 instead of 2038
on 32-bit architectures.

Hopefully, everybody has already moved to a newer format long
ago (v2 was introduced with linux-2.4), but it's hard to be sure.

Change-Id: I1b50e16b4cbf7ecdae690de6d372f1cc06cb1e09
Signed-off-by: Arnd Bergmann <arnd@arndb.de>
Git-commit: 0ee09cbf5222e8ba62175fc88d8d802cc6bb32c2
Git-repo: git://git.kernel.org/pub/scm/linux/kernel/git/arnd/playground.git
Signed-off-by: Parikshit Pareek <quic_ppareek@quicinc.com>