y2038: itimer: change implementation to timespec64
There is no 64-bit version of getitimer/setitimer since that is not
actually needed. However, the implementation is built around the
deprecated 'struct timeval' type.

Change the code to use timespec64 internally to reduce the dependencies
on timeval and associated helper functions.

Minor adjustments in the code are needed to make the native and compat
version work the same way, and to keep the range check working after
the conversion.

Change-Id: I5e469b74aa1d01b3717fd7a244150b52edc3090f
Signed-off-by: Arnd Bergmann <arnd@arndb.de>
Git-commit: 5aa5b3877da002d3b9eb96ec06bbac8102262a7e
Git-repo: git://git.kernel.org/pub/scm/linux/kernel/git/arnd/playground.git
Signed-off-by: Parikshit Pareek <quic_ppareek@quicinc.com>