mtd: msm_qpic_nand: check for page_erased bit along with op_err
Due to a hardware bug in ECC-Engine, ECC-Engine couldn't able
to set OP_ERR bit in flash_status regesiter whenever an erased
page encounters bitflips. Due to this, ECC-Engine is trying to
correct the bitflips on an erased page and leading to data
corruption. So, a check for PAGE_ERASED bit is added prior to
OP_ERR bit for an erased page detection logic.

Change-Id: I570625123fe828450dade06570f782ebe93d39f1
Signed-off-by: Pradeep P V K <ppvk@codeaurora.org>
Signed-off-by: Pradeep P V K <pragalla@qti.qualcomm.com>