usb: gadget: f_uac1: Fix unsupported control request handling
Currently, the driver returns positive value, either request len
or sizeof(int), even if the control selector ID is not found in
the list of supported controls by the driver/device. This leads
to garbage data being returned by the device, instead of expected
STALL. Fix this by properly handling the INTERFACE ctrl requests
and returning -EOPNOTSUPP if function does not support it.

Change-Id: I5c4de195a678dbd370ffba4e16851e827b5536fb
Signed-off-by: Ajay Agarwal <ajaya@codeaurora.org>