cfg80211/nl80211: Kernel support to enable SAE and OWE
This interface allows the host driver to offload the authentication to
user space. This can be used to implement SAE without full
implementation in the kernel/firmware while still being able to use
NL80211_CMD_CONNECT with driver-based BSS selection.

This interface allows the host driver to offload OWE processing to
user space. This intends to support OWE/Enhanced Open AKM by the host
drivers that implement SME but rely on the user space for
the crypotographic/OWE processing.Such drivers are not capable of
processing/deriving the DH IE's.

Change-Id: Ica7b05625e76c270117a702d6f76a9c94d8eda50
Signed-off-by: Wei Yang <yangwei@codeaurora.org>
Signed-off-by: Long Cheng <quic_longchen@quicinc.com>