mm/memblock.c: fix bug in early_dyn_memhotplug
memblock.memory.cnt will decrease if we remove all memory
blocks in a region. We will lose the next region in this
condition. The solution is we save memblock.memory.cnt
before we scan through a region and compare the new
counter with the old one after that, we decrease the idx
if we find counter changes. This way, we are always using
the correct idx before we enter next loop.

Change-Id: I11d31eb14e51a2d1c0e2f587911bddd540a75ccd
Signed-off-by: Jiacheng Zheng <jiaczhen@codeaurora.org>